feat(android): VideoControllerConfiguration.scale
