+++ media_kit_video/lib/media_kit_video.dart
@@ -6,6 +6,7 @@
 
 export 'package:media_kit_video/src/video/video.dart';
 export 'package:media_kit_video/src/video_controller/video_controller.dart';
+export 'package:media_kit_video/src/video_controller/platform_video_controller.dart';
 
 export 'package:media_kit_video/src/subtitle/subtitle_view.dart';
