+++ media_kit_video/lib/src/video_controller/native_video_controller/stub.dart
@@ -5,7 +5,6 @@
 /// Use of this source code is governed by MIT license that can be found in the LICENSE file.
 import 'package:media_kit/media_kit.dart';
 
-import 'package:media_kit_video/src/video_controller/video_controller.dart';
 import 'package:media_kit_video/src/video_controller/platform_video_controller.dart';
 
 // Stub declaration for avoiding compilation errors on Dart JS using conditional imports.
